102773: Adjustments reported by Resharper.
