+++ src/Equinor.ProCoSys.IPO.Command/InvitationCommands/IpoSendMailException.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Transactions;
 
 namespace Equinor.ProCoSys.IPO.Command.InvitationCommands
 {
